Refactor: Join => User 로 통합
