--- src/main/java/com/example/hobbyheavy/controller/JoinController.java
@@ -1,37 +1,0 @@
-package com.example.hobbyheavy.controller;
-
-import com.example.hobbyheavy.dto.request.JoinRequest;
-import com.example.hobbyheavy.exception.CustomException;
-import com.example.hobbyheavy.service.JoinService;
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
-
-@Slf4j
-@RestController
-@RequiredArgsConstructor
-public class JoinController {
-
-    private final JoinService joinService;
-
-    @PostMapping("/join")
-    public ResponseEntity<String> joinProcess(@Valid @RequestBody JoinRequest joinRequest) {
-        log.info("회원가입 요청: {}", joinRequest);
-        try {
-            joinService.joinProcess(joinRequest);
-            log.info("회원가입 처리 완료: {}", joinRequest.getUserId());
-            return ResponseEntity.status(HttpStatus.CREATED).body("회원가입 성공");
-        } catch (CustomException e) {
-            log.error("회원가입 중 오류 발생: {}", e.getMessage());
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(e.getMessage());
-        } catch (Exception e) {
-            log.error("예상치 못한 오류 발생: ", e);
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("회원가입 중 오류가 발생했습니다.");
-        }
-    }
-}
