+++ src/main/java/com/example/hobbyheavy/controller/UserController.java
@@ -3,27 +3,48 @@ package com.example.hobbyheavy.controller;
 import com.example.hobbyheavy.auth.CustomUserDetails;
 import com.example.hobbyheavy.dto.request.DeleteUserRequest;
 import com.example.hobbyheavy.dto.request.PasswordUpdateRequest;
+import com.example.hobbyheavy.dto.request.JoinRequest;
 import com.example.hobbyheavy.dto.response.UserInfoResponse;
 import com.example.hobbyheavy.exception.CustomException;
 import com.example.hobbyheavy.exception.ExceptionCode;
 import com.example.hobbyheavy.service.UserService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 
+@Slf4j
 @RestController
-@RequestMapping("/user")
+@RequestMapping
 @RequiredArgsConstructor
 public class UserController {
 
     private final UserService userService;
 
+    /** 회원가입 **/
+    @PostMapping("/join")
+    public ResponseEntity<String> joinUser(@Valid @RequestBody JoinRequest joinRequest) {
+        log.info("회원가입 요청: {}", joinRequest);
+        try {
+            userService.JoinUser(joinRequest);
+            log.info("회원가입 처리 완료: {}", joinRequest.getUserId());
+            return ResponseEntity.status(HttpStatus.CREATED).body("회원가입 성공");
+        } catch (CustomException e) {
+            log.error("회원가입 중 오류 발생: {}", e.getMessage());
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(e.getMessage());
+        } catch (Exception e) {
+            log.error("예상치 못한 오류 발생: ", e);
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("회원가입 중 오류가 발생했습니다.");
+        }
+    }
+
     /** 나의 회원정보 조회 **/
-    @GetMapping("/my-info")
+    @GetMapping("/user/my-info")
     public ResponseEntity<UserInfoResponse> getMyInfo(@AuthenticationPrincipal UserDetails userDetails) {
 
         // userDetails에서 userId 추출
@@ -35,7 +56,7 @@ public class UserController {
     }
 
     /** 비밀번호 변경 **/
-    @PutMapping("/password")
+    @PutMapping("/user/password")
     public ResponseEntity<String> updatePassword(@Valid @RequestBody PasswordUpdateRequest request, Authentication authentication) {
 
         // 현재 인증된 사용자 ID를 가져옴
@@ -50,7 +71,7 @@ public class UserController {
     /* 비밀번호 찾기 **/
 
     /** 회원 탈퇴 **/
-    @DeleteMapping
+    @DeleteMapping("/user")
     public ResponseEntity<String> deleteUser(@Valid @RequestBody DeleteUserRequest request, @AuthenticationPrincipal CustomUserDetails userDetails) {
         String password = request.getPassword();
